requirements.txt
absl-py
astunparse
blinker
certifi
charset-normalizer
click
filelock
Flask
flatbuffers
fsspec
gast
google-pasta
grpcio
gunicorn
h5py
huggingface-hub
idna
itsdangerous
Jinja2
joblib
keras
libclang
Markdown
markdown-it-py
MarkupSafe
mdurl
ml-dtypes
namex
nltk
numpy
opencv-python
opt_einsum
optree
packaging
pillow
protobuf
Pygments
PyYAML
regex
requests
rich
safetensors
scikit-learn
scipy
setuptools
six
tensorboard
tensorboard-data-server
tensorflow
termcolor
threadpoolctl
tokenizers
tqdm
transformers
typing_extensions
urllib3
Werkzeug
wheel
wrapt

